+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The major version bump is due to dropped support for PHP `8.1` and a several bre
 
 * Mako applications are now better suited to run on application servers like FrankenPHP.
 * The development error view now displays the Mako environment name.
+* Added `Connection::blob()` method that allows you to easily fetch a blob column as stream.
+* Added `Query::blob()` method that allows you to easily fetch a blob column as stream.
 
 #### Improvements